+++ src/main/java/gift/Application.java
@@ -1,11 +1,9 @@
 package gift;
 
 
-import gift.domain.Product;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.filter.HiddenHttpMethodFilter;
 
 @SpringBootApplication
