+++ src/main/java/gift/model/user/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.model.user;
 
 public class AuthenticationRequest {
     private String email;
