+++ src/main/java/gift/repository/user/UserRepository.java
@@ -1,10 +1,10 @@
 package gift.repository.user;
 
-import gift.domain.User;
+import gift.model.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    public User findAllByEmail(String email);
+    User findAllByEmail(String email);
 }
